// Frontend/src/api/models/UserSummary.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * User summary for token response
 */
export type UserSummary = {
    id: number;
    username: string;
    email: string;
    is_superuser: boolean;
    department?: (string | null);
    roles?: Array<string>;
};